+++ core-ng/src/main/java/core/framework/api/module/SystemModule.java
@@ -36,6 +36,7 @@ public final class SystemModule extends Module {
 
         property("sys.log.traceLogPath").ifPresent(path -> log().traceLogPath(Paths.get(path)));
         property("sys.log.actionLogPath").ifPresent(path -> log().actionLogPath(Paths.get(path)));
+        property("sys.log.remoteLogHost").ifPresent(host -> log().remoteLogHost(host));
 
         property("sys.rabbitMQ.host").ifPresent(hosts -> queue().rabbitMQ().hosts(hosts.split(",")));
         property("sys.rabbitMQ.user").ifPresent(user -> queue().rabbitMQ().user(user));