+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQ.java
@@ -1,13 +1,9 @@
 package core.framework.impl.queue;
 
-import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.Address;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
-import core.framework.api.log.ActionLogContext;
-import core.framework.api.util.StopWatch;
-import core.framework.api.util.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -45,7 +41,6 @@ public class RabbitMQ {
     }
 
     public void hosts(String... hosts) {
-        logger.info("set rabbitMQ hosts, hosts={}", Arrays.toString(hosts));
         addresses = new Address[hosts.length];
         for (int i = 0; i < hosts.length; i++) {
             String host = hosts[i];
@@ -58,27 +53,15 @@ public class RabbitMQ {
     }
 
     public RabbitMQConsumer consumer(String queue, int prefetchCount) {
-        Channel channel = channel();
+        Channel channel = createChannel();
         return new RabbitMQConsumer(channel, queue, prefetchCount);
     }
 
-    public void publish(String exchange, String routingKey, String message, AMQP.BasicProperties properties) {
-        StopWatch watch = new StopWatch();
-        Channel channel = null;
-        try {
-            channel = channel();
-            channel.basicPublish(exchange, routingKey, properties, Strings.bytes(message));
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        } finally {
-            closeChannel(channel);
-            long elapsedTime = watch.elapsedTime();
-            ActionLogContext.track("rabbitMQ", elapsedTime);
-            logger.debug("publish, exchange={}, routingKey={}, messageId={}, elapsedTime={}", exchange, routingKey, properties.getMessageId(), elapsedTime);
-        }
+    public RabbitMQChannel channel() {
+        return new RabbitMQChannel(createChannel());
     }
 
-    private Channel channel() {
+    private Channel createChannel() {
         try {
             synchronized (this) {
                 if (connection == null) {
@@ -88,16 +71,6 @@ public class RabbitMQ {
             return connection.createChannel();
         } catch (IOException | TimeoutException e) {
             throw new Error(e);
-        }
-    }
-
-    private void closeChannel(Channel channel) {
-        if (channel != null) {
-            try {
-                channel.close();
-            } catch (IOException | TimeoutException e) {
-                throw new Error(e);
-            }
         }
     }
 }