+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQPublisher.java
@@ -52,7 +52,9 @@ public class RabbitMQPublisher<T> implements MessagePublisher<T> {
 
         linkContext(builder, headers);
 
-        rabbitMQ.publish(exchange, routingKey, JSON.toJSON(message), builder.build());
+        try (RabbitMQChannel channel = rabbitMQ.channel()) {
+            channel.publish(exchange, routingKey, JSON.toJSON(message), builder.build());
+        }
     }
 
     private void linkContext(AMQP.BasicProperties.Builder builder, Map<String, Object> headers) {