+++ core-ng/src/main/java/core/framework/impl/scheduler/JobTrigger.java
@@ -1,0 +1,17 @@
+package core.framework.impl.scheduler;
+
+import core.framework.api.scheduler.Job;
+import core.framework.impl.scheduler.trigger.Trigger;
+
+/**
+ * @author neo
+ */
+class JobTrigger {
+    final Job job;
+    final Trigger trigger;
+
+    JobTrigger(Job job, Trigger trigger) {
+        this.job = job;
+        this.trigger = trigger;
+    }
+}