+++ log-processor/src/main/java/Main.java
@@ -1,8 +1,10 @@
+import core.log.LogProcessorApp;
+
 /**
  * @author neo
  */
 public class Main {
     public static void main(String[] args) {
-
+        new LogProcessorApp().start();
     }
 }